requirements.txt
fastapi
uvicorn
python-dotenv
mysql-connector-python
scikit-learn
qiskit
qiskit-machine-learning
